+++ packages/provider-utils/src/types/tool-call.ts
@@ -1,6 +1,6 @@
 /**
-Typed tool call that is returned by generateText and streamText. 
-It contains the tool call ID, the tool name, and the tool arguments. 
+Typed tool call that is returned by generateText and streamText.
+It contains the tool call ID, the tool name, and the tool arguments.
  */
 export interface ToolCall<NAME extends string, ARGS> {
   /**
